+++ tests/test_checks.py
@@ -22,7 +22,7 @@ def test_onchain_and_expected_transfer_equality():
         account_id="egld-ping-pong",
         bech32="erd1qqqqqqqqqqqqqpgqpxkd9qgyyxykq5l6d8v9zud99hpwh7l0plcq3dae77",
         saved_values={"PingAmount": 1000000000000000000},
-        wasm_hash="1383133d22b8be01c4dc6dfda448dbf0b70ba1acb348a50dd3224b9c8bb21757",
+        code_hash="1383133d22b8be01c4dc6dfda448dbf0b70ba1acb348a50dd3224b9c8bb21757",
         deploy_time=1677261606,
         last_upgrade_time=1677261606,
     )